add latest ios offline code
