+++ ios-offline-maps/ios-offline-maps/ios_offline_mapsApp.swift
@@ -2,13 +2,16 @@
 //  ios_offline_mapsApp.swift
 //  ios-offline-maps
 //
-//  Created by Chris Whong on 8/13/25.
+//  Created by Chris Whong on 9/17/25.
 //
 
 import SwiftUI
 
 @main
 struct ios_offline_mapsApp: App {
+    init() {
+        OfflineRegionManager.ensureStylePackDownloaded()
+    }
     var body: some Scene {
         WindowGroup {
             ContentView()
